remove unused GatewayForms and move them to form utils
