+++ src/pages/namespace/Explorer/Endpoint/EndpointEditor/Form/components/FormHelper.tsx
@@ -1,5 +1,9 @@
-import { FC, PropsWithChildren } from "react";
+import { FC, PropsWithChildren, useState } from "react";
+import { Plus, X } from "lucide-react";
 
+import Button from "~/design/Button";
+import { ButtonBar } from "~/design/ButtonBar";
+import Input from "~/design/Input";
 import { twMergeClsx } from "~/util/helpers";
 
 type FieldsetProps = PropsWithChildren & {
@@ -29,3 +33,108 @@ export const Fieldset: FC<FieldsetProps> = ({
     {children}
   </fieldset>
 );
+
+type ArrayInputProps = {
+  placeholder?: string;
+  externalArray: string[];
+  onChange: (newValue: string[]) => void;
+};
+
+export const ArrayInput: FC<ArrayInputProps> = ({
+  externalArray,
+  onChange,
+  placeholder,
+}) => {
+  const [internalArray, setInternalArray] = useState(externalArray);
+  const [inputVal, setInputVal] = useState("");
+
+  const newValue = (val: string) => {
+    if (val.length) {
+      setInternalArray((old) => {
+        const newValue = [...old, inputVal];
+        const newValueRemovedEmpty = newValue.filter(Boolean);
+        onChange(newValueRemovedEmpty);
+        setInputVal("");
+        return newValueRemovedEmpty;
+      });
+    }
+  };
+
+  const changeValue = (valueIndex: number, newValue: string) => {
+    setInternalArray((oldArray) => {
+      const newArray = oldArray.map((oldValue, index) => {
+        if (index === valueIndex) {
+          return newValue;
+        }
+        return oldValue;
+      });
+
+      if (newValue) {
+        onChange(newArray);
+      }
+      return newArray;
+    });
+  };
+
+  const deleteValue = (valueIndex: number) => {
+    setInternalArray((old) => {
+      const newValue = old.filter((_, i) => i !== valueIndex);
+      const newValueRemovedEmpty = newValue.filter(Boolean);
+      onChange(newValueRemovedEmpty);
+      return newValueRemovedEmpty;
+    });
+  };
+
+  return (
+    <div className="grid grid-cols-2 gap-5">
+      {internalArray.map((value, valueIndex) => (
+        <ButtonBar key={valueIndex}>
+          <Input
+            placeholder={placeholder}
+            value={value}
+            onChange={(e) => {
+              changeValue(valueIndex, e.target.value);
+            }}
+          />
+          {}
+          <Button
+            icon
+            variant="outline"
+            type="button"
+            onClick={() => {
+              deleteValue(valueIndex);
+            }}
+          >
+            <X />
+          </Button>
+        </ButtonBar>
+      ))}
+
+      <ButtonBar>
+        <Input
+          placeholder={placeholder}
+          value={inputVal}
+          onChange={(e) => {
+            setInputVal(e.target.value);
+          }}
+          onKeyDown={(e) => {
+            if (e.key === "Enter") {
+              newValue(inputVal);
+              e.preventDefault();
+            }
+          }}
+        />
+        <Button
+          icon
+          variant={!inputVal ? "outline" : undefined}
+          onClick={() => {
+            newValue(inputVal);
+          }}
+          type="button"
+        >
+          <Plus />
+        </Button>
+      </ButtonBar>
+    </div>
+  );
+};
